Working deletion of topics if user is the creator.

+++ CloudQuestions_web/questions/views.py
@@ -45,8 +45,14 @@ def index(request):
 
 def detail(request, topic):
     global QUESTION_LIST
+    breakpoint()
     questions_by_topic = question_service.questions_by_topic(topic)
     color = None
+    context = {}
+    context['creator'] = False
+    creator = question_service.get_creator(topic).id
+    if creator == request.user.id:
+        context['creator'] = True
     if request.GET.get('red_button') == 'Bad':
         color = 'red'
     elif request.GET.get('yellow_button') == 'Medium':
@@ -56,16 +62,17 @@ def detail(request, topic):
     if color:
         question_service.update_stats(topic, color, request.user)
         return redirect('accounts:settings', topic, color)
-    if request.GET.get('delete'):
-        topic_id = question_service.topics_by_name(topic)[0]
+    if request.GET.get('delete') and context['creator']:
+        topic_id = question_service.topics_by_id(topic)[0]
         topic = Topic.objects.get(id=topic_id)
         topic.delete()
+        return redirect('questions:index')
     if request.GET.get('random'):
         QUESTION_LIST.clear()
         return redirect('questions:random', topic)
-    return render(request, 'questions/detail.html',
-                  {'topic': topic,
-                   'questions_by_topic': questions_by_topic})
+    context['topic'] = topic
+    context['questions_by_topic'] = questions_by_topic
+    return render(request, 'questions/detail.html', context)
 
 
 def random_questions(request, topic):
